pyproject.toml
[project]
name = "ai-devs4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
